Update config tests
